Remove GNUism from Makefile
I suspect that the 'export SSHPASS = ...' was a GNU/Linuxism in GMake
and I'd like to keep this Makefile portable for BSD :) Unfortunately,
the next easiest thing I could think of was either to prepend
SSHPASS=... to the start of every sshpass(1) command or to export it
first and then escape every newline to keep all the commands in the same
subshell and same environment. I think the latter was the best of the
two.